Added Match Item. Hit and toggle light. Document of how to add item
Added item offset in itemdata for placing item in hand.
Added collision height in itemdata for picking up items
Added a document of how to add an item to the game.
Added a match item. Can hit and turn flame on and off.
Candle can also turn on and off flame.
Moved flame down with offset so bottom of flame stays on object
Updated collectable to change collision shape based on itemdata
Added a functions for activating items in hand.